renaming the exa_common
