+++ exaudfclient/exa_udfclient.cc
@@ -1,4 +1,4 @@
-#include "exa_common/exa_udf_client.h"
+#include "exa_udfclient_binary_common/exa_udf_client.h"
 
 int main(int argc, char **argv) {
     ExaUdfClient client;
